# docs/openapi.json
{"openapi":"3.0.1","info":{"title":"vui_quiz_backend API","description":"API documentation for vui_quiz_backend","termsOfService":"http://swagger.io/terms/","license":{"name":"Apache 2.0","url":"http://springdoc.org"},"version":"v1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Quiz Management API","description":"Endpoints for creating, retrieving, and managing quizzes"}],"paths":{"/api/session/join/{gamePin}":{"post":{"tags":["game-session-controller"],"operationId":"joinSession","parameters":[{"name":"gamePin","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/api/session/create":{"post":{"tags":["game-session-controller"],"operationId":"createSession","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/api/quizzes":{"post":{"tags":["Quiz Management API"],"summary":"Create a new quiz","description":"Creates a new quiz including its questions. The response contains quiz metadata; questions are not re-fetched in this response.","operationId":"createQuiz","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizDTO"}}},"required":true},"responses":{"201":{"description":"Quiz created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizDTO"}}}}},"security":[{"bearerAuth":[]}]}},"/api/auth/signup":{"post":{"tags":["auth-controller"],"summary":"Register a new user","description":"Creates a new user account.","operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequest"}}},"required":true},"responses":{"400":{"description":"Invalid request data (e.g., username/email already exists)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"200":{"description":"User registered successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}}}}},"/api/auth/signin":{"post":{"tags":["auth-controller"],"summary":"Authenticate user and get JWT","description":"Provides a JWT token upon successful authentication.","operationId":"authenticateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"400":{"description":"Invalid request payload","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"200":{"description":"Successfully authenticated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JwtResponse"}}}},"401":{"description":"Unauthorized - Invalid credentials","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}}}}},"/api/quizzes/{quizId}":{"get":{"tags":["Quiz Management API"],"summary":"Get full quiz details by ID","description":"Retrieves a specific quiz along with all its questions and their details.","operationId":"getQuizDetailsById","parameters":[{"name":"quizId","in":"path","description":"ID of the quiz to retrieve","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"404":{"description":"Quiz not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/QuizDTO"}}}},"200":{"description":"Quiz details found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizDTO"}}}}}}},"/api/quizzes/public":{"get":{"tags":["Quiz Management API"],"summary":"Get public and published quizzes","description":"Retrieves a paginated list of quizzes that are marked as public (visibility=1) and published (status='PUBLISHED').","operationId":"getPublicPublishedQuizzes","parameters":[{"name":"pageable","in":"query","description":"Pagination and sorting parameters (e.g., page=0&size=10&sort=createdAt,desc)","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"List of public quizzes retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Page"}}}}}}},"/api/quizzes/my-quizzes":{"get":{"tags":["Quiz Management API"],"summary":"Get quizzes created by the current user","description":"Retrieves a paginated list of quizzes created by the currently authenticated user.","operationId":"getCurrentUserQuizzes","parameters":[{"name":"pageable","in":"query","description":"Pagination and sorting parameters (e.g., page=0&size=10&sort=modifiedAt,desc)","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"401":{"description":"Unauthorized - User not logged in","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageQuizDTO"}}}},"200":{"description":"List of user's quizzes retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Page"}}}}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"ChoiceDTO":{"required":["correct"],"type":"object","properties":{"answer":{"type":"string","description":"Text content of the answer choice. Used if 'image' is not present for this choice.","example":"True"},"image":{"$ref":"#/components/schemas/ImageDetailDTO"},"correct":{"type":"boolean","description":"Indicates if this choice is a correct answer."}},"description":"Represents an answer choice for a question"},"ImageDetailDTO":{"type":"object","properties":{"url":{"type":"string","description":"Direct URL (file_path) to the image. This is provided in requests and responses.","example":"https://placehold.co/200x200/blue/white?text=Square"},"altText":{"type":"string","description":"Alternative text for the image","example":"A Square"},"contentType":{"type":"string","description":"Content type of the image. Optional, backend might infer or retrieve if image is internal.","example":"image/png"}},"description":"Details of an image, often used within question choices. The URL is the primary field."},"QuestionDTO":{"required":["position","title","type"],"type":"object","properties":{"id":{"type":"string","description":"Unique identifier for the question (generated by backend for response)","format":"uuid"},"type":{"type":"string","description":"Type of the question block","example":"quiz","enum":["content","quiz","jumble","survey","open_ended"]},"title":{"type":"string","description":"The question text, or title for 'content' type slides. Can contain HTML.","example":"Which technology is NOT part of the frontend stack?"},"description":{"type":"string","description":"Detailed description, primarily for 'content' type slides.","example":"This section will test your knowledge..."},"time":{"type":"integer","description":"Time limit for the question in milliseconds. Not applicable for 'content' type.","format":"int32","example":30000},"pointsMultiplier":{"type":"integer","description":"Points multiplier for the question. Not applicable for 'content' or 'survey' types.","format":"int32","example":1},"choices":{"type":"array","description":"List of choices for the question. Structure depends on the question type. Not applicable for 'content' type where it can be null/empty.","items":{"$ref":"#/components/schemas/ChoiceDTO"}},"image":{"type":"string","description":"URL (file_path) of the main image associated with the question. Provided in requests and responses.","example":"https://placehold.co/600x400/purple/white?text=4-Choice+Quiz"},"video":{"$ref":"#/components/schemas/VideoDetailDTO"},"media":{"type":"array","description":"Array for additional media elements (e.g., URLs or structured objects). For now, simple strings.","example":[],"items":{"type":"string","description":"Array for additional media elements (e.g., URLs or structured objects). For now, simple strings.","example":"[]"}},"position":{"type":"integer","description":"0-based order/position of this question within the quiz.","format":"int32","example":1}},"description":"Represents a single question or content slide within a quiz"},"QuizDTO":{"required":["questions","title","visibility"],"type":"object","properties":{"visibility":{"type":"integer","description":"Visibility status (e.g., 1 for public, 0 for private)","format":"int32","example":1},"title":{"type":"string","description":"Quiz title","example":"Sample Quiz - All Question Types"},"description":{"type":"string","description":"Quiz description","example":"This quiz contains one example of each question type for frontend mocking purposes."},"quizType":{"type":"string","description":"General type of the quiz, from mock's 'quizType'. Consider if this maps to a Quiz 'status' or a dedicated field.","example":"quiz"},"questions":{"type":"array","description":"Array of Question Objects","items":{"$ref":"#/components/schemas/QuestionDTO"}},"questionCount":{"type":"integer","description":"The number of questions in this quiz. Populated by the backend.","format":"int32","example":10},"isValid":{"type":"boolean","description":"Server-side validation flag (populated by backend for responses)","example":true},"playAsGuest":{"type":"boolean","description":"Flag indicating if guests can play (game session property, not quiz typically)","example":true},"status":{"type":"string","description":"Status of the quiz (e.g., DRAFT, PUBLISHED). If not provided in request, defaults to DRAFT.","example":"DRAFT"},"type":{"type":"string","description":"Redundant type field from mock? Matches mock's 'type' field.","example":"quiz"},"created":{"type":"integer","description":"Creation timestamp (Unix millis, populated by backend for responses)","format":"int64","example":1678886400000},"modified":{"type":"integer","description":"Last modified timestamp (Unix millis, populated by backend for responses)","format":"int64","example":1745286000000},"tags":{"type":"array","description":"List of tag names associated with the quiz. Provided during creation, included in responses.","example":["Science","2019","Trivia"],"items":{"type":"string","description":"List of tag names associated with the quiz. Provided during creation, included in responses.","example":"[\"Science\",\"2019\",\"Trivia\"]"}},"uuid":{"type":"string","description":"Unique Quiz Identifier (generated by backend for responses)","format":"uuid"},"creator":{"type":"string","description":"User ID of the creator (set by backend based on authenticated user)","format":"uuid","example":"39567e72-f833-4bcc-80b0-0c2b02d6c921"},"creator_username":{"type":"string","description":"Username of the creator (populated by backend for responses)","example":"MockUser"},"cover":{"type":"string","description":"URL (file_path) of the cover image for the quiz. Provided in requests and responses.","example":"https://placehold.co/600x400/orange/white?text=Sample+Quiz+Cover"},"lobby_video":{"$ref":"#/components/schemas/VideoDetailDTO"}},"description":"Represents a Quiz, used for both creation requests and responses"},"VideoDetailDTO":{"type":"object","properties":{"id":{"type":"string","description":"Video ID from the service (e.g., YouTube video ID)","example":"dQw4w9WgXcQ"},"startTime":{"type":"number","description":"Start time for the video playback in seconds","format":"double","example":0.0},"endTime":{"type":"number","description":"End time for the video playback in seconds","format":"double","example":0.0},"service":{"type":"string","description":"Video service provider","example":"youtube"},"fullUrl":{"type":"string","description":"Full URL to the video (can be derived if ID and service are known)","example":"https://www.youtube.com/watch?v=dQw4w9WgXcQ"}},"description":"Details for a video (e.g., YouTube)"},"SignupRequest":{"required":["password","username"],"type":"object","properties":{"username":{"maxLength":50,"minLength":3,"type":"string","description":"Desired username for the new account","example":"newuser"},"email":{"maxLength":200,"minLength":0,"type":"string","description":"Email address for the new account","example":"newuser@example.com"},"password":{"maxLength":30,"minLength":6,"type":"string","description":"Password for the new account","example":"securePassword123"}},"description":"Request payload for user registration"},"MessageResponse":{"type":"object","properties":{"message":{"type":"string"}}},"LoginRequest":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string","description":"Username of the user","example":"testuser"},"password":{"type":"string","description":"Password of the user","example":"password123"}}},"JwtResponse":{"type":"object","properties":{"token":{"type":"string"},"type":{"type":"string"},"id":{"type":"string","format":"uuid"},"username":{"type":"string"},"email":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"Page":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"type":"object"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"paged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"direction":{"type":"string"},"nullHandling":{"type":"string"},"ascending":{"type":"boolean"},"property":{"type":"string"},"ignoreCase":{"type":"boolean"}}},"PageQuizDTO":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/QuizDTO"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"Enter JWT Bearer token **_only_**","name":"bearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}