# src/main/java/com/vuiquiz/quizwebsocket/controller/QuizController.java
package com.vuiquiz.quizwebsocket.controller;

import com.vuiquiz.quizwebsocket.dto.QuizDTO;
import com.vuiquiz.quizwebsocket.security.services.UserDetailsImpl;
import com.vuiquiz.quizwebsocket.service.QuizService;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.data.domain.Sort;
import org.springframework.data.web.PageableDefault;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.web.bind.annotation.*;

import java.util.UUID;

@CrossOrigin(origins = "*", maxAge = 3600)
@RestController
@RequestMapping("/api/quizzes")
@Tag(name = "Quiz Management API", description = "Endpoints for creating, retrieving, and managing quizzes")
public class QuizController {

    private final QuizService quizService;

    @Autowired
    public QuizController(QuizService quizService) {
        this.quizService = quizService;
    }

    @PostMapping
    @PreAuthorize("isAuthenticated()")
    @Operation(summary = "Create a new quiz",
            description = "Creates a new quiz including its questions. The response contains quiz metadata; questions are not re-fetched in this response.",
            security = @SecurityRequirement(name = "bearerAuth"))
    @ApiResponse(responseCode = "201", description = "Quiz created successfully",
            content = @Content(mediaType = "application/json", schema = @Schema(implementation = QuizDTO.class)))
    // ... other ApiResponse annotations
    public ResponseEntity<QuizDTO> createQuiz(@Valid @RequestBody QuizDTO quizRequestDTO) {
        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
        UserDetailsImpl userDetails = (UserDetailsImpl) authentication.getPrincipal();
        UUID creatorId = userDetails.getId();

        QuizDTO createdQuiz = quizService.createQuiz(quizRequestDTO, creatorId);
        return new ResponseEntity<>(createdQuiz, HttpStatus.CREATED);
    }

    @GetMapping("/{quizId}")
    @Operation(summary = "Get full quiz details by ID",
            description = "Retrieves a specific quiz along with all its questions and their details.")
    @ApiResponse(responseCode = "200", description = "Quiz details found",
            content = @Content(mediaType = "application/json", schema = @Schema(implementation = QuizDTO.class)))
    @ApiResponse(responseCode = "404", description = "Quiz not found")
    public ResponseEntity<QuizDTO> getQuizDetailsById( // Renamed controller method for clarity too, though not strictly necessary
                                                       @Parameter(description = "ID of the quiz to retrieve") @PathVariable UUID quizId) {
        QuizDTO quizDTO = quizService.getQuizDetailsById(quizId); // Call the renamed service method
        return ResponseEntity.ok(quizDTO);
    }

    @GetMapping("/my-quizzes")
    @PreAuthorize("isAuthenticated()") // User must be logged in
    @Operation(summary = "Get quizzes created by the current user",
            description = "Retrieves a paginated list of quizzes created by the currently authenticated user.")
    @ApiResponse(responseCode = "200", description = "List of user's quizzes retrieved successfully",
            content = @Content(mediaType = "application/json", schema = @Schema(implementation = Page.class)))
    @ApiResponse(responseCode = "401", description = "Unauthorized - User not logged in")
    @SecurityRequirement(name = "bearerAuth") // Indicates JWT is needed
    public ResponseEntity<Page<QuizDTO>> getCurrentUserQuizzes(
            @Parameter(description = "Pagination and sorting parameters (e.g., page=0&size=10&sort=modifiedAt,desc)")
            @PageableDefault(
                    size = 10,
                    sort = "modifiedAt", // Specify property name here
                    direction = Sort.Direction.DESC // Specify direction separately
            ) Pageable pageable) { // Default sort by modified date
        Page<QuizDTO> userQuizzes = quizService.getQuizzesByCurrentUser(pageable);
        return ResponseEntity.ok(userQuizzes);
    }

    @GetMapping("/public")
    @Operation(summary = "Get public and published quizzes",
            description = "Retrieves a paginated list of quizzes that are marked as public (visibility=1) and published (status='PUBLISHED').")
    @ApiResponse(responseCode = "200", description = "List of public quizzes retrieved successfully",
            content = @Content(mediaType = "application/json", schema = @Schema(implementation = Page.class)))
    // Response is a Page
    public ResponseEntity<Page<QuizDTO>> getPublicPublishedQuizzes(
            @Parameter(description = "Pagination and sorting parameters (e.g., page=0&size=10&sort=createdAt,desc)")
            @PageableDefault(size = 10, sort = "createdAt") Pageable pageable) { // Inject Pageable
        Page<QuizDTO> publicQuizzes = quizService.getPublicPublishedQuizzes(pageable);
        return ResponseEntity.ok(publicQuizzes);
    }
}